fix: CI/CD move tests pyarrow module
