+++ tests/pipeline/test_pipeline_extra.py
@@ -22,6 +22,7 @@ import dlt
 from dlt.common import json, pendulum
 from dlt.common.destination import DestinationCapabilitiesContext
 from dlt.common.destination.capabilities import TLoaderFileFormat
+from dlt.destinations.impl.filesystem.filesystem import FilesystemClient
 from dlt.common.runtime.collector import (
     AliveCollector,
     EnlightenCollector,
@@ -599,3 +600,82 @@ def test_pick_matching_file_format(test_storage: FileStorage) -> None:
     files = test_storage.list_folder_files("user_data_csv/object")
     assert len(files) == 1
     assert files[0].endswith("csv")
+
+
+def test_filesystem_column_hint_timezone() -> None:
+    import pyarrow.parquet as pq
+    import posixpath
+
+    os.environ["DESTINATION__FILESYSTEM__BUCKET_URL"] = "_storage"
+
+    # talbe: events_timezone_off
+    @dlt.resource(
+        columns={"event_tstamp": {"data_type": "timestamp", "timezone": False}},
+        primary_key="event_id",
+    )
+    def events_timezone_off():
+        yield [
+            {"event_id": 1, "event_tstamp": "2024-07-30T10:00:00.123+00:00"},
+            {"event_id": 2, "event_tstamp": "2024-07-30T10:00:00.123456+02:00"},
+            {"event_id": 3, "event_tstamp": "2024-07-30T10:00:00.123456"},
+        ]
+
+    # talbe: events_timezone_on
+    @dlt.resource(
+        columns={"event_tstamp": {"data_type": "timestamp", "timezone": True}},
+        primary_key="event_id",
+    )
+    def events_timezone_on():
+        yield [
+            {"event_id": 1, "event_tstamp": "2024-07-30T10:00:00.123+00:00"},
+            {"event_id": 2, "event_tstamp": "2024-07-30T10:00:00.123456+02:00"},
+            {"event_id": 3, "event_tstamp": "2024-07-30T10:00:00.123456"},
+        ]
+
+    # talbe: events_timezone_unset
+    @dlt.resource(
+        primary_key="event_id",
+    )
+    def events_timezone_unset():
+        yield [
+            {"event_id": 1, "event_tstamp": "2024-07-30T10:00:00.123+00:00"},
+            {"event_id": 2, "event_tstamp": "2024-07-30T10:00:00.123456+02:00"},
+            {"event_id": 3, "event_tstamp": "2024-07-30T10:00:00.123456"},
+        ]
+
+    pipeline = dlt.pipeline(destination="filesystem")
+
+    pipeline.run(
+        [events_timezone_off(), events_timezone_on(), events_timezone_unset()],
+        loader_file_format="parquet",
+    )
+
+    client: FilesystemClient = pipeline.destination_client()  # type: ignore[assignment]
+
+    expected_results = {
+        "events_timezone_off": None,
+        "events_timezone_on": "UTC",
+        "events_timezone_unset": "UTC",
+    }
+
+    for t in expected_results.keys():
+        events_glob = posixpath.join(client.dataset_path, f"{t}/*")
+        events_files = client.fs_client.glob(events_glob)
+
+        with open(events_files[0], "rb") as f:
+            table = pq.read_table(f)
+
+            # convert the timestamps to strings
+            timestamps = [
+                ts.as_py().strftime("%Y-%m-%dT%H:%M:%S.%f") for ts in table.column("event_tstamp")
+            ]
+            assert timestamps == [
+                "2024-07-30T10:00:00.123000",
+                "2024-07-30T08:00:00.123456",
+                "2024-07-30T10:00:00.123456",
+            ]
+
+            # check if the Parquet file contains timezone information
+            schema = table.schema
+            field = schema.field("event_tstamp")
+            assert field.type.tz == expected_results[t]
